updated gatsby version and all dependencies
