+++ gatsby-config.js
@@ -1,10 +1,17 @@
-// const path = require(`path`)
-
+/**
+ * @type {import('gatsby').GatsbyConfig}
+ */
 module.exports = {
+  siteMetadata: {
+    title: `swfl-coders-website`,
+    siteUrl: `https://www.yourdomain.tld`
+  },
   plugins: [
     {
-      resolve: `gatsby-plugin-material-ui`,
-      options: {},
+      resolve: 'gatsby-plugin-manifest',
+      options: {
+        "icon": "src/images/icon.png"
+      }
     },
     {
       resolve: "gatsby-plugin-react-svg",
