+++ package.json
@@ -1,47 +1,41 @@
 {
-  "name": "swfl-coders",
+  "name": "swfl-coders-website",
+  "version": "1.0.0",
   "private": true,
-  "description": "A central hub for all things related to the SWFL Coders meetup group and SWFL technology",
-  "version": "0.1.0",
-  "license": "MIT",
+  "description": "swfl-coders-website",
+  "author": "PBH",
+  "keywords": [
+    "gatsby"
+  ],
   "scripts": {
-    "build": "gatsby build",
     "develop": "gatsby develop",
-    "format": "prettier --write src/**/*.{js,jsx}",
-    "start": "npm run develop",
+    "start": "gatsby develop",
+    "build": "gatsby build",
     "serve": "gatsby serve",
-    "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\""
+    "clean": "gatsby clean"
   },
   "dependencies": {
-    "@material-ui/core": "^4.2.1",
-    "@material-ui/icons": "^4.2.1",
-    "@material-ui/styles": "^4.2.1",
-    "fetch-jsonp": "^1.1.3",
-    "gatsby": "^2.13.25",
-    "gatsby-image": "^2.2.7",
-    "gatsby-plugin-material-ui": "^2.1.4",
-    "gatsby-plugin-react-helmet": "^3.3.14",
-    "gatsby-plugin-react-svg": "^2.1.1",
-    "gatsby-plugin-sass": "^2.1.13",
-    "gatsby-plugin-sharp": "^2.2.9",
-    "gatsby-source-filesystem": "^2.1.7",
-    "gatsby-transformer-remark": "^2.6.10",
-    "gatsby-transformer-sharp": "^2.2.5",
-    "markdown-to-jsx": "^6.10.2",
-    "moment": "^2.24.0",
-    "node-sass": "^4.12.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.14.13",
+    "@mui/styles": "^5.14.13",
+    "fetch-jsonp": "^1.3.0",
+    "gatsby": "^5.12.4",
+    "gatsby-image": "^3.11.0",
+    "gatsby-plugin-manifest": "^5.12.1",
+    "gatsby-plugin-react-helmet": "^6.12.0",
+    "gatsby-plugin-react-svg": "^3.3.0",
+    "gatsby-plugin-sass": "^6.12.1",
+    "gatsby-plugin-sharp": "^5.12.1",
+    "gatsby-source-filesystem": "^5.12.0",
+    "gatsby-transformer-remark": "^6.12.1",
+    "gatsby-transformer-sharp": "^5.12.1",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0"
   },
   "devDependencies": {
-    "prettier": "^1.18.2"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/gatsbyjs/gatsby-starter-hello-world"
-  },
-  "bugs": {
-    "url": "https://github.com/gatsbyjs/gatsby/issues"
+    "prettier": "^3.0.3"
   }
 }
