+++ src/components/company_card_preview.js
@@ -1,11 +1,11 @@
 import React from "react"
-import { makeStyles } from "@material-ui/core/styles"
-import Card from "@material-ui/core/Card"
-import CardActionArea from "@material-ui/core/CardActionArea"
-import CardContent from "@material-ui/core/CardContent"
-import Typography from "@material-ui/core/Typography"
-import CardMedia from "@material-ui/core/CardMedia"
-import Hidden from "@material-ui/core/Hidden"
+import { makeStyles } from "@mui/styles"
+import Card from "@mui/material/Card"
+import CardActionArea from "@mui/material/CardActionArea"
+import CardContent from "@mui/material/CardContent"
+import Typography from "@mui/material/Typography"
+import CardMedia from "@mui/material/CardMedia"
+import Hidden from "@mui/material/Hidden"
 
 const useStyles = makeStyles(theme => ({
   cardArea: {
