+++ src/components/header.js
@@ -1,6 +1,6 @@
 import React from "react"
-import { makeStyles } from "@material-ui/core/styles"
-import Toolbar from "@material-ui/core/Toolbar"
+import { makeStyles } from "@mui/styles"
+import Toolbar from "@mui/material/Toolbar"
 import SocialMedia from "./social_media"
 import CodersLogo from "../assets/svg-logos/coders-logo.svg"
 import { Helmet } from "react-helmet"
