+++ src/components/meetup_groups.js
@@ -1,12 +1,12 @@
 import React from "react"
-import { makeStyles } from "@material-ui/core/styles"
-import Grid from "@material-ui/core/Grid"
-import Typography from "@material-ui/core/Typography"
-import Card from "@material-ui/core/Card"
-import CardActionArea from "@material-ui/core/CardActionArea"
-import CardContent from "@material-ui/core/CardContent"
-import CardMedia from "@material-ui/core/CardMedia"
-import Hidden from "@material-ui/core/Hidden"
+import { makeStyles } from "@mui/styles"
+import Grid from "@mui/material/Grid"
+import Typography from "@mui/material/Typography"
+import Card from "@mui/material/Card"
+import CardActionArea from "@mui/material/CardActionArea"
+import CardContent from "@mui/material/CardContent"
+import CardMedia from "@mui/material/CardMedia"
+import Hidden from "@mui/material/Hidden"
 
 const meetupGroups = [
   {
