+++ src/components/nav.js
@@ -1,7 +1,7 @@
 import React from "react"
-import Toolbar from "@material-ui/core/Toolbar"
-import AppBar from "@material-ui/core/AppBar"
-import Button from "@material-ui/core/Button"
+import Toolbar from "@mui/material/Toolbar"
+import AppBar from "@mui/material/AppBar"
+import Button from "@mui/material/Button"
 
 const links = [
   { label: "Home", path: "/" },
