+++ src/components/social_media.js
@@ -1,6 +1,6 @@
 import React from "react"
-import Link from "@material-ui/core/Link"
-import { makeStyles } from "@material-ui/core/styles"
+import Link from "@mui/material/Link"
+import { makeStyles } from "@mui/styles"
 import MeetupIcon from "../assets/svg-logos/meetup-brands.svg"
 import SlackIcon from "../assets/svg-logos/slack-brands.svg"
 import FacebookIcon from "../assets/svg-logos/facebook-f-brands.svg"
